+++ lib/domain/commands/place_component_command.dart
@@ -2,7 +2,7 @@ import 'package:circuitquest/domain/commands/command.dart';
 import 'package:circuitquest/core/components/base/component.dart';
 import 'package:circuitquest/domain/models/placed_component.dart';
 import 'package:flutter/material.dart';
-import 'package:circuitquest/domain/sandbox/sandbox_engine.dart';
+import 'package:circuitquest/domain/use_cases/sandbox_engine.dart';
 
 /// Command to place a component on the canvas.
 class PlaceComponentCommand extends Command {
