+++ lib/domain/commands/remove_component_command.dart
@@ -1,7 +1,7 @@
 import 'package:circuitquest/domain/commands/command.dart';
 import 'package:circuitquest/domain/models/placed_component.dart';
 import 'package:circuitquest/domain/models/wire_connection.dart';
-import 'package:circuitquest/domain/sandbox/sandbox_engine.dart';
+import 'package:circuitquest/domain/use_cases/sandbox_engine.dart';
 
 /// Command to remove a component from the canvas.
 class RemoveComponentCommand extends Command {
