+++ lib/domain/use_cases/level_validation/level_validation_result.dart
