--- lib/levels/levels.dart
@@ -1,24 +1,0 @@
-/// Level system library
-/// 
-/// This library provides all necessary classes and utilities for working with
-/// the CircuitQuest level system.
-/// 
-/// Main classes:
-/// - [Level] - Data model for a level
-/// - [LevelLoader] - Service for loading levels from JSON assets
-/// - [LevelUtils] - Utility functions for level operations
-/// 
-/// Usage:
-/// ```dart
-/// import 'package:circuitquest/levels/levels.dart';
-/// 
-/// final loader = LevelLoader();
-/// final level = await loader.loadLevel(0);
-/// ```
-
-library levels;
-
-export 'level.dart';
-export '../data/services/level_loader_service.dart';
-export 'level_utils.dart';
-export 'level_validator.dart';
