+++ lib/ui/level_mode/level_screen.dart
@@ -1,9 +1,9 @@
+import 'package:circuitquest/domain/models/level.dart';
 import 'package:circuitquest/l10n/app_localizations.dart';
 import 'package:circuitquest/ui/level_mode/level_bottom_app_bar.dart';
 import 'package:circuitquest/ui/level_mode/level_component_palette.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:circuitquest/levels/levels.dart';
 import 'package:circuitquest/constants.dart';
 import '../shared/widgets/circuit_canvas/circuit_canvas.dart';
 import '../shared/widgets/control_panel.dart';
